week6/pyproject.toml
[project]
name = "week6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.11.0",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
]
